Rework of models' repo structure
