+++ ovulation_predicting/models/__init__.py
@@ -1,0 +1,16 @@
+__all__ = [
+    'ClassificationMLP',
+    'CnnLstm',
+    'FeedBack',
+    'NoisySinCurve',
+    'WideCnn',
+    'MyModelWrapper'
+]
+
+
+from .model_baseline import NoisySinCurve
+from .model_classification import ClassificationMLP
+from .model_cnn import WideCnn
+from .model_cnn_lstm import CnnLstm
+from .models import MyModelWrapper
+from .model_rnn import FeedBack
